Add admin, economy, and stash systems
This commit adds admin command support, queued database writes, heartbeat payload handling, and economy/stash management to the server. It also updates client packet handling, fixes damage notification logic in script data, and adds protocol tests for heartbeat, chat, snapshot, and damage packets.

+++ zone-server/internal/network/session.go
@@ -16,7 +16,7 @@ type PlayerSession struct {
 	Rotation          [2]float32
 	Velocity          [3]float32
 	Health            float32
-	AnimFlags         uint16
+	AnimFlags         uint8
 	InSafeZone        bool
 	SafeZoneID        string
 	LastSeen          time.Time
